+++ io/shared/src/main/scala/fs2/io/net/SocketGroup.scala
@@ -24,7 +24,8 @@ package io
 package net
 
 import cats.effect.kernel.Resource
-import com.comcast.ip4s.{Host, IpAddress, Port, SocketAddress}
+import cats.syntax.all._
+import com.comcast.ip4s.{Host, IpAddress, Ipv4Address, Port, SocketAddress}
 import cats.effect.kernel.Async
 
 /** Supports creation of client and server TCP sockets that all share
@@ -71,23 +72,32 @@ trait SocketGroup[F[_]] {
   ): Resource[F, (SocketAddress[IpAddress], Stream[F, Socket[F]])]
 }
 
-private[net] object SocketGroup extends SocketGroupCompanionPlatform {
+private[net] object SocketGroup {
 
-  private[net] abstract class AbstractAsyncSocketGroup[F[_]: Async] extends SocketGroup[F] {
-    def server(
-        address: Option[Host],
-        port: Option[Port],
-        options: List[SocketOption]
-    ): Stream[F, Socket[F]] =
-      Stream
-        .resource(
-          serverResource(
-            address,
-            port,
+  def fromIpSockets[F[_]: Async](ipSockets: IpSocketsProvider[F]): SocketGroup[F] =
+    new SocketGroup[F] {
+      def client(to: SocketAddress[Host], options: List[SocketOption]) =
+        ipSockets.connect(to, options)
+
+      def server(
+          address: Option[Host],
+          port: Option[Port],
+          options: List[SocketOption]
+      ): Stream[F, Socket[F]] =
+        Stream.resource(serverResource(address, port, options)).flatMap(_._2)
+
+      def serverResource(
+          address: Option[Host],
+          port: Option[Port],
+          options: List[SocketOption]
+      ): Resource[F, (SocketAddress[IpAddress], Stream[F, Socket[F]])] =
+        ipSockets
+          .bind(
+            SocketAddress(address.getOrElse(Ipv4Address.Wildcard), port.getOrElse(Port.Wildcard)),
             options
           )
-        )
-        .flatMap { case (_, clients) => clients }
-  }
-
+          .evalMap(b =>
+            b.localAddressGen.map(_.asInstanceOf[SocketAddress[IpAddress]]).tupleRight(b.accept)
+          )
+    }
 }
